requirements.txt
Pillow==5.3.0
piexif==1.0.13
selenium==3.141.0
